# api/api_documentation.yaml
openapi: 3.0.0
info:
  title: AI Coordinator API
  description: >
    This API allows interaction with the AI Coordinator system. It supports operations for users, data processing, and log retrieval.
  version: "1.0.0"
paths:
  /users/:
    post:
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
      responses:
        200:
          description: User created successfully
        400:
          description: User already exists
        500:
          description: Server error
  /upload/:
    post:
      summary: Upload and process a file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        200:
          description: Data processed successfully
        500:
          description: Server error
  /logs/:
    get:
      summary: Retrieve API logs
      responses:
        200:
          description: Logs retrieved successfully
        500:
          description: Server error
  /api/memory/generate-text/:
    post:
      summary: Generate text using OpenAI GPT
      description: >
        This endpoint uses OpenAI GPT model to generate text based on the provided prompt.
        You can control the length of the generated text using the `max_tokens` parameter.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                prompt:
                  type: string
                  description: The input text for the GPT model.
                  example: "Hello, AI! What is the weather like today?"
                max_tokens:
                  type: integer
                  description: The maximum number of tokens to generate in the response.
                  example: 150
      responses:
        200:
          description: Successfully generated text
          content:
            application/json:
              schema:
                type: object
                properties:
                  generated_text:
                    type: string
                    description: The generated text from GPT.
        500:
          description: Error during text generation
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    description: Error message